Django admin page fix
Django admin page would not display all modules. Folder structure
changed to reflect 1.8 changes, and code changed to reflect deprecated
tags.
