--- agentex/templates/admin/base_site_old.html
@@ -1,9 +1,0 @@
-{% extends "admin/base.html" %}
-
-{% block title %}{{ title }} | {{ site_title|default:_('LCOGT') }}{% endblock %}
-
-{% block branding %}
-<h1 id="site-name"><a href="{% url 'admin:index' %}">{{ site_header|default:_('Django administration') }}</a></h1>
-{% endblock %}
-
-{% block nav-global %}{% endblock %}
