+++ agentex/templates/agentex/agentex_info.html
@@ -20,7 +20,7 @@
        <tr><th colspan=2>Completed planets</th></tr>
    {% for p in completes.planets %}
    <tr class="{% cycle 'row1' 'row2' %}">
-       <td><a href="{% url agentex_all_calib p.id %}">{{p.name}}</a></td><td>{{p.subtotal}}</td>
+       <td><a href="{% url 'agentex_all_calib' p.id %}">{{p.name}}</a></td><td>{{p.subtotal}}</td>
    </tr>
    {% endfor %}
    </tbody>
