+++ agentex/views.py
@@ -47,6 +47,8 @@ import agentex.dataset as ds
 from django.conf import settings
 from agentex.agentex_settings import planet_level
 
+from agentex.datareduc import calibrator_data, average_combine, calibrator_averages, photometry
+
 # Added by TJ to allow logged in query to function (bottom of document)
 #from django.contrib.sessions.models import Session
 
@@ -482,245 +484,6 @@ def addvalue(request,code):
     else:
         result = addvalue_nopost(request, person, code)
         return result
-    
-'''
-    #####################################################################################
-    #####################################################################################
-
-    # If statement to check if form posts
-    #if (request.POST):
-    def addvalue_post(request, code):
-    ####### Form data has been submitted
-        x = []
-        y = []
-        nocals = request.POST.get('calibrators','1')
-        setting = request.POST.get('entrytype','')
-        # Only update the user's preference if they change it
-        if (setting == 'manual' and o[0].dataexploreview == True):
-            o.update(dataexploreview=False)
-            messages.success(request, "Setting changed to use manual view")
-            entrymode = 'M'
-        elif (setting == 'dataexplorer' and o[0].dataexploreview == False):
-            o.update(dataexploreview=True)
-            messages.success(request, "Setting changed to use web view")
-            entrymode = 'W'
-        else:
-            entrymode = 'N'
-        id = request.POST.get('dataid','')
-        form = DataEntryForm(request.POST)
-        if form.is_valid():
-            cd = form.cleaned_data
-            ind = {'source':'S','bg':'B'}
-            reduced = 0
-            update = request.POST.get('update','')
-            counts = list()
-            for i in ind:
-                value = float(cd[i+'counts'])
-                x.append(cd[i+'xpos'])
-                y.append(cd[i+'ypos'])
-                counts.append(value)
-            for vari in range(1,int(nocals)+1):
-                cali = str(vari)
-                value = request.POST.get('cal'+cali+'counts','')
-                x.append(request.POST.get('cal'+cali+'xpos',''))
-                y.append(request.POST.get('cal'+cali+'ypos',' '))
-                counts.append(float(value))
-            pointsum = {'bg' :  '%.2f' % counts[0], 'sc' : '%.2f' % counts[1], 'cal' : counts[2:]}
-            if (len(x) < 3 or len(y) < 3):
-                messages.warning(request,'Please submit calibration, blank sky and source apertures.')
-                url = reverse('agentex.views.addvalue',args= [DataSource.objects.get(id=id).event.name])
-                return HttpResponseRedirect(url)
-            x = map(float,x)
-            y = map(float,y)
-            coords = zip(x,y)
-            dataid = request.POST.get('dataid','')
-            resp = savemeasurement(person,pointsum,coords,dataid,entrymode)
-            messages.add_message(request, resp['code'], resp['msg'])
-            if webin == False:
-                url = "%s?%s" % (reverse('agentex.views.addvalue',args= [DataSource.objects.get(id=id).event.name]),"input=manual")
-            else:
-                url = reverse('agentex.views.addvalue',args= [DataSource.objects.get(id=id).event.name])
-             #messages.success(request, "Measurement successfully added")
-            return HttpResponseRedirect(url)
-        else:
-            return render_to_response('agentex/dataentry.html', {'data':DataSource.objects.get(id=id),'form':form,'data_url':settings.DATA_URL}, context_instance=RequestContext(request))
-    
-    #####################################################################################
-    #####################################################################################
-
-    #else:
-    def addvalue_nopost:
-        nextcal = request.GET.get('next',False)
-        ############ This condition is active when a user edits the frame
-        # Find the data sources for the given code
-        source = DataSource.objects.filter(event__name=code)
-        length = source.count()
-        # Pull out data user has viewed and exclude them from the list of possible candidates
-        ds = Datapoint.objects.values_list('data',flat=True).filter(data__event__name=code,user=person,pointtype='S')
-        input = request.GET.get('input',False)
-        id = request.GET.get('dataid',False)
-        # If an ID is specified return the frame, as long as the person has made measurements of it
-        if id:
-            dnext = False
-            #### If anonymous user tell them they cannot edit points
-            if person == guestuser:
-                messages.warning(request,'You cannot edit points unless you are logged in')
-                try:
-                    url = reverse('agentex.views.addvalue',args= [DataSource.objects.get(id=id).event.name])
-                    return HttpResponseRedirect(url)
-                except:
-                    raise Http404
-            mycalibs = []
-            ##### The page is being displayed with data for editing
-            points = Datapoint.objects.filter(data__id=id,user=person)
-            if nextcal=='cal':
-                print "*******"
-                dp = Datapoint.objects.filter(pointtype='S',user=person,data__id=id)
-                dd = dp[0].data.timestamp
-                ds = Datapoint.objects.filter(pointtype='S',user=person,data__timestamp__gt=dd).order_by('data__timestamp')
-                if ds.count() > 0:
-                    dnext = ds[0].data
-            d = DataSource.objects.filter(id=id)[0]
-            otherpoints = Datapoint.objects.filter(~Q(user=person),pointtype='C',data=d)
-            cals = Datapoint.objects.values_list('xpos','ypos','radius').filter(data=d,pointtype='C').order_by('coorder__calid')
-            calibs = []
-            if cals:
-                for c in cals:
-                    calibs.append({'x' : int(c[0]) , 'y' : int(c[1]), 'r' : int(c[2])})
-            source = points.filter(pointtype='S')[:1]
-            bg = points.filter(pointtype='B')[:1]
-            #### If there are no results, the person is hacking the query string. Return a fresh frame
-            if (source.count() == 0 or bg.count() == 0):
-                url = reverse('agentex.views.addvalue',args= [code])
-                return HttpResponseRedirect(url)
-            cal = points.filter(pointtype='C').order_by('coorder__calid')
-            for c in cal:
-                line = {'x' :c.xpos,'y' : c.ypos}
-                mycalibs.append(line)
-            ### If more cals have been placed on other frames add these to this frame
-            max_cal = Datapoint.objects.filter(pointtype='C',user=person).aggregate(max=Max('coorder__calid'))['max']
-            if max_cal+1 > cal.count():
-                for order in range(cal.count(),max_cal+1):
-                    dp = Datapoint.objects.filter(pointtype='C',user=person,coorder__calid=order,data__event__name=code)
-                    if dp.count() > 0:
-                        line = {'x': dp[0].xpos, 'y':dp[0].ypos}
-                        # Add to the mycalibs array
-                        mycalibs.append(line)
-            coords = { 'source': {'x' :source[0].xpos,'y' : source[0].ypos},
-                     'cal'  : mycalibs,
-                     'bg'  : {'x' :bg[0].xpos,'y' : bg[0].ypos},
-                     'radius' : source[0].radius,
-                     'id'  : id,
-                     'numcals' : len(mycalibs),
-                     }
-            messages.info(request, "Updating measurement")
-            return render_to_response('agentex/dataentry.html',{'data':d,
-                                                                    'next':dnext,
-                                                                    'points':coords,
-                                                                    'update':True,
-                                                                    'webinput':webin,
-                                                                    'progress':progress,
-                                                                    'form':form,
-                                                                    'calibrators':calibs,
-                                                                    'least_data':least_coords,
-                                                                    'data_url':settings.DATA_URL},
-                                    context_instance=RequestContext(request))                             
-        else:
-            ######## User is being given a new frame not editing data  
-            complete = 0
-            if  (progress['done'] >= progress['total'] and person != guestuser):
-                ####### No new data can be provided because the user has come to the end
-                complete = 1
-                numplanets = DataCollection.objects.values('planet').filter(person=person,complete=True).annotate(Count('complete')).count()
-                e = Event.objects.filter(name=code)[0]
-                resp = achievementscheck(person,e,0,0,0,0,numplanets)
-
-                msg = '<br />'
-                for item in resp:
-                    if messages.SUCCESS == item['code'] :
-                        msg += "<img src=\""+STATIC_URL+item['image']+"\" style=\"width:96px;height:96px;\" alt=\"Badge\" />"
-                        messages.success(request,msg)
-                
-                return HttpResponseRedirect(reverse('my-graph',args=[code]))
-
-                return render_to_response('agentex/dataentry.html',
-                                        {'event': e,
-                                        'complete':complete,
-                                        'progress':progress,
-                                        'points':Datapoint.objects.filter(user=person,pointtype='S',data__event=e).order_by('data__timestamp'),
-                                        'data_url':settings.DATA_URL,
-                                        'numplanets':numplanets,},
-                                        context_instance=RequestContext(request))            
-            else:
-                planet = Event.objects.get(name=code)
-                mylist = Datapoint.objects.filter(user=person,pointtype='S',data__event=planet).values_list('data',flat=True)
-                print mylist
-                ### if person does not have a DataCollection it is their first measurement
-                if (DataCollection.objects.filter(planet=planet,person=person).count() == 0):
-                    d = DataSource.objects.filter(event=planet,id=planet.finder)[0]
-                    did = d.id
-                    try:
-                        dold = d.id
-                        first = True   
-                    except:
-                        messages.error(request,"Finderchart cannot be found")
-                        raise Http404    
-                elif  person == guestuser:
-                    d = DataSource.objects.filter(event=planet).annotate(count=Count('datapoint')).order_by('-count')[0]
-                    did = d.id
-                    dold = d.id
-                    first = True
-                else:
-                    try:
-                        source_rank = DataSource.objects.filter(event=planet ).annotate(count=Count('datapoint') ).values_list('id','count').order_by('-count')  
-                        available = [x for x in source_rank if x[0] not in list(mylist)]
-                        dold = Datapoint.objects.values_list('data__id',flat=True).filter(user=person,data__event=planet,pointtype='C').annotate(max =Max('coorder__calid')).order_by('-max','-taken')[0]
-                    # Find position in set of DataSources
-                        d = available[0]
-                        did = d[0]
-                        first = False
-                    except Exception,e:
-                        print e
-                        messages.error(request,"User has a data collection but no points!")
-                        raise Http404
-                cals = Datapoint.objects.values_list('xpos','ypos').filter(data=dold,pointtype='C',user=person).order_by('coorder__calid')
-                calibs = []
-                if cals:
-                    for c in cals:
-                        calibs.append({'x' : int(c[0]) , 'y' : int(c[1])})
-                otherpoints = Datapoint.objects.filter(~Q(user=person),pointtype='C',data__id=did)
-                othercals = []
-                if otherpoints:
-                    for c in otherpoints:
-                        othercals.append({'x' : int(c.xpos) , 'y' : int(c.ypos),'r':int(c.radius)})
-                prev = Datapoint.objects.filter(user=person,data=dold).order_by('coorder__calid')
-                if first == False:
-                    coords = { 'source': {'x' :prev.filter(pointtype='S')[0].xpos,'y' : prev.filter(pointtype='S')[0].ypos},
-                             'bg'  : {'x' :prev.filter(pointtype='B')[0].xpos,'y' : prev.filter(pointtype='B')[0].ypos},
-                             'cal'  : calibs ,
-                             'id'  : dold,
-                             'radius' : planet.radius
-                             }
-                else:
-                    coords = False
-                if person == guestuser:
-                    progress = {'percent'   : "0",
-                                'done'      : 0,
-                                'total'     : n_sources,}
-                return render_to_response('agentex/dataentry.html',
-                                        {'data':DataSource.objects.get(id=did),
-                                        'complete':complete,
-                                        'update':False,
-                                        'webinput':webin,
-                                        'progress':progress,
-                                        'form':form,
-                                        'calibrators':othercals,
-                                        'points':coords,
-                                        'least_data':least_coords,
-                                        'data_url':settings.DATA_URL},
-                                        context_instance=RequestContext(request))        
-
-'''
 
 def savemeasurement(person,pointsum,coords,dataid,entrymode):
     # Only update the user's preference if they change it
@@ -1050,30 +813,7 @@ def my_data(o,code):
                 }
         data.append(line)
     return data,points
-    
-def calibrator_data(calid,code):
-    data = []
-    sources, times = zip(*DataSource.objects.filter(event__name=code).values_list('id','timestamp').order_by('timestamp'))
-    points  = Datapoint.objects.filter(data__in=sources)
-    #points.filter(pointtype='C').values('data__id','user','value')
-    people = Decision.objects.filter(source__id=calid,planet__name=code,value='D',current=True).values_list('person__username',flat=True).distinct()
-    norm = dict((key,0) for key in sources)
-    for pid in people:
-        cal = []
-        sc = dict(points.filter(user__username=pid,pointtype='S').values_list('data__id','value'))
-        bg = dict(points.filter(user__username=pid,pointtype='B').values_list('data__id','value'))
-        c = dict(points.filter(user__username=pid,pointtype='C',coorder__source__id=calid).values_list('data__id','value'))
-        sc_norm = dict(norm.items() + sc.items())
-        bg_norm = dict(norm.items() + bg.items())
-        c_norm = dict(norm.items() + c.items())
-        #print sc_norm,bg_norm,c_norm
-        for v in sources:
-            try:
-                cal.append((sc_norm[v]- bg_norm[v])/(c_norm[v] - bg_norm[v]))
-            except:
-                cal.append(0)
-        data.append(cal)
-    return data,[timegm(s.timetuple())+1e-6*s.microsecond for s in times],list(people)
+
         
 @login_required  
 def graphview(request,code,mode,calid):
@@ -1595,83 +1335,9 @@ def myaverages(code,person):
     # If they have no 'D' decisions
     return [0.]*ds.count()
     
-def calibrator_averages(code,person=None,progress=False):
-    cals = []
-    cats = []
-    planet = Event.objects.get(name=code)
-    sources = list(DataSource.objects.filter(event=planet).order_by('timestamp').values_list('id','timestamp'))
-    ids,stamps = zip(*sources)
-    if person:
-        ## select calibrator stars used, excluding ones where ID == None, i.e. non-catalogue stars
-        dc = DataCollection.objects.filter(~Q(source=None),person=person,planet=planet).order_by('calid')
-        ## Measurement values only for selected 'person'
-        dps = Datapoint.objects.filter(data__event=planet,user=person).order_by('data__timestamp')
-    else:
-        # select calibrator stars used, excluding ones where ID == None, i.e. non-catalogue stars
-        dc = DataCollection.objects.filter(~Q(source=None),planet=planet).order_by('calid')
-        ## Measurement values only for selected 'person'
-        dps = Datapoint.objects.filter(data__event=planet).order_by('data__timestamp')
-    averages = AverageSet.objects.filter(planet=planet)
-    if person:
-        # Make a combined list of source values
-        measurements = dps.filter(pointtype='S')
-        sc = average_combine(measurements,averages,ids,None,'S',progress)
-        # Make a combined list of background values
-        measurements = dps.filter(pointtype='B')
-        bg = average_combine(measurements,averages,ids,None,'B',progress)
-    else:
-        sc = array(averages.filter(star=None,settype='S')[0].data)
-        bg = array(averages.filter(star=None,settype='B')[0].data)
-    # Make a combined list of all calibration stars used by 'person'
-    for calibrator in dc:
-        if person:
-            measurements = dps.filter(pointtype='C',coorder=calibrator)
-            ave = average_combine(measurements,averages,ids,calibrator.source,'C',progress)
-        else:
-            ave_cal = averages.filter(star=calibrator,settype='C')
-            if ave_cal.count() > 0:
-                ave = array(ave_cal[0].data)
-            else:
-                ave = array([])
-        if ave.size > 0:
-            cals.append(ave)
-            try:
-                if person:
-                    decvalue = Decision.objects.filter(source=calibrator.source,person=person,planet=planet,current=True)[0].value
-                else:
-                    decvalue = Decision.objects.filter(source=calibrator.source, planet=planet,current=True)[0].value
-            except:
-                decvalue ='X'
-            cat_item = {'sourcename':calibrator.source.name,'catalogue':calibrator.source.catalogue}
-            cat_item['decsion'] = decvalue
-            cat_item['order'] = str(calibrator.calid)
-            cats.append(cat_item)
-    return cals,sc,bg,stamps,ids,cats
+
     
-def average_combine(measurements,averages,ids,star,category,progress,admin=False):
-    if progress['done'] < progress['total']:
-        ave_measurement = averages.filter(star=star,settype=category)
-        if ave_measurement.count() > 0:
-            ## Find the array indices of my values and replace these averages
-            ave = array(ave_measurement[0].data)
-            mine = zip(*measurements.values_list('data','value'))
-            try:
-                my_ids = [ids.index(x) for x in mine[0]]
-                ave[my_ids] = mine[1]
-            except Exception, e:
-                print e
-            return ave
-        else:
-            return array([])
-    elif progress['done'] == progress['total']:
-        mine = array(measurements.values_list('value',flat=True))
-        return mine
-    elif not progress:
-        print "No progress was passed"
-        return array([])
-    else:
-        print "Error - too many measurements: %s %s" % (measurements.count() , numobs)
-        return array([])
+
         
 def admin_averagecals(code,person):
     # Uses and SQL statement to try to speed up the query for averaging data points
@@ -1788,41 +1454,7 @@ def admin_averagecals(code,person):
         return normcals,stamps,[],[]
     return cals,normcals,[],[],dates,stamps,[],cats
     
-def photometry(code,person,progress=False,admin=False):
-    
-    # Empty lists to store normalised calibrators and maximum values
-    normcals = []
-    maxvals = []
 
-    # Call in averages
-    cals,sc,bg,times,ids,cats = calibrator_averages(code,person,progress)
-    
-    indexes = [int(i) for i in ids]
-    #sc = array(sc)
-    #bg = array(bg)     
-    
-    # Iterate over every calibrator
-    for cal in cals:
-	if len(cal) == progress['total']:
-        #### Do not attempt to do the photmetry where the number of calibrators does not match the total        
-            # Determine calibrated flux from source
-            val = (sc - bg)/(cal-bg)
-            # Determine maximum flux from source
-            maxval = mean(r_[val[:3],val[-3:]])
-            # Append to maxvals
-            maxvals.append(maxval)
-            # Normalise the maxval
-            norm = val/maxval
-            #Append the normalised value
-            normcals.append(list(norm))
-        # Find my data and create unix timestamps
-    unixt = lambda x: timegm(x.timetuple())+1e-6*x.microsecond
-    iso = lambda x: x.isoformat(" ")
-    stamps = map(unixt,times)
-    dates = map(iso,times)
-    if admin:
-        return normcals,stamps,indexes,cats
-    return cals,normcals,list(sc),list(bg),dates,stamps,indexes,cats
         
 def averagecals_async(e):
     #e = Event.objects.get(name=code)
