+++ templates/admin/base_site.html
@@ -1,0 +1,3 @@
+{% extends "admin/base.html" %}
+
+{% block title %}{{ title }} | {{ site_title|default:_('LCOGT') }}{% endblock %}
